Correção no register no frontend

+++ Frontend/src/pages/Profile/index.js
@@ -27,22 +27,26 @@ export default function Profile() {
     }, [ongId]);
 
     async function handleDeleteIncident(id) {
-        try {
-            await api.delete(`incidents/${id}`, {
-                headers: {
-                    Authorization: ongId,
-                }
-            });
-            setIncidents(incidents.filter(incident => incident.id !== id))
-        } catch (err) {
-            alert('Erro ao deletar caso, tente novamente.');
+        if (window.confirm("Deseja excluir realmente ?")) {
+            try {
+                await api.delete(`incidents/${id}`, {
+                    headers: {
+                        Authorization: ongId,
+                    }
+                });
+                setIncidents(incidents.filter(incident => incident.id !== id))
+            } catch (err) {
+                alert('Erro ao deletar caso, tente novamente.');
+            }
         }
     }
 
     function handleLogout() {
-        localStorage.clear();
+        if (window.confirm("Deseja sair ?")) {
+            localStorage.clear();
 
-        history.push('/');
+            history.push('/');
+        }
     }
 
     return (
